requirements.txt
rembg
Pillow
tk
